GearZone/gearzone/requirements.txt
amqp==2.5.2
aniso8601==7.0.0
autopep8==1.4.4
babel==2.7.0
beautifulsoup4==4.7.1
billiard==3.6.1.0
bleach==2.1.4
boto3==1.10.8
botocore==1.13.8
braintree==3.49.0
cachetools==3.1.1
djangorestframework

cairocffi==1.1.0
cairosvg==2.4.2
celery==4.3.0
django-allauth == 0.39.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
dataclasses==0.6; python_version < "3.7"
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==0.2.0
django==2.2.9
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap4==1.0.1
django-cache-url==3.0.0
django-countries==5.5
django-elasticsearch-dsl==0.5.1
django-filter==2.2.0
django-graphql-jwt==0.2.1
django-impersonate==1.4.1
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.10.0
django-phonenumber-field==2.4.0
django-prices==2.1.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.6
django-silk==2.0.0
django-storages==1.7.2
django-templated-email==2.3.0
django-versatileimagefield==1.11
django-webpack-loader==0.6.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
elasticsearch==6.3.1
elasticsearch-dsl==6.0.1
faker==2.0.4
freezegun==0.3.12
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gprof2dot==2016.10.13
graphene==2.1.8
graphene-django==2.6.0
graphene-django-optimizer==0.6.1
graphene-federation==0.0.3
graphql-core==2.2.1
graphql-relay==2.0.0
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
ipaddress==1.0.23
jinja2==2.10.3
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.6
lxml==4.4.1
markdown==3.1.1
markupsafe==1.1.1
maxminddb==1.5.1
maxminddb-geolite2==2018.703
measurement==3.1.0
more-itertools==7.2.0
mpmath==1.1.0
oauthlib==3.1.0
phonenumberslite==8.10.23
pillow==6.2.1
prices==1.0.0
promise==2.2.1
protobuf==3.10.0
psycopg2-binary==2.8.4
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pygments==2.4.2
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.8.1
python-magic==0.4.15
python-magic-bin==0.4.14; sys_platform == "win32"
python3-openid==3.1.0; python_version >= "3.0"
pytz==2019.3
razorpay==1.2.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
rx==1.6.1
s3transfer==0.2.1
sentry-sdk==0.13.2
singledispatch==3.4.0.3
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
stripe==2.40.0
sympy==1.4
text-unidecode==1.3
tinycss2==1.0.2
tqdm==4.38.0
typing==3.7.4.1
urllib3==1.25.7
uwsgi==2.0.18; sys_platform != "win32"
vine==1.3.0
weasyprint==50
webencodings==0.5.1
zipp==0.6.0
pandas==0.25.3
openpyxl==3.0.0
